requirements.txt
Django==4.2
Unidecode
django-social-auth
#social-auth-app-django-mongoengine
httplib2
oauth2
python-openid
# wsgiref py2.7
django-gravatar2
django-extensions
markdown2
pymongo
django-markitup
Markdown
djangorestframework
django-filter
django-redis
redis
requests
nltk
textblob
anora
snowballstemmer
typogrify
networkx
psycopg2-binary
jinja2
python-social-auth
